Create story body component
