+++ src/components/StoryView.jsx
@@ -2,9 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Image from './StoryView/Header/Image';
 import Contents from './StoryView/Header/Contents';
+import Body from './StoryView/Body';
 import Style from './StoryView/Header/Style';
 
-const StoryView = ({ title, primaryImage }) => (
+const StoryView = ({
+  title,
+  teaser,
+  body,
+  primaryImage,
+}) => (
   <div className="story-wrapper">
     <Style />
     <div className="story-header">
@@ -12,15 +18,19 @@ const StoryView = ({ title, primaryImage }) => (
         <Contents title={title} />
       </Image>
     </div>
+    <Body teaser={teaser} body={body} />
   </div>
 );
 
 StoryView.defaultProps = {
+  teaser: '',
   primaryImage: {},
 };
 
 StoryView.propTypes = {
   title: PropTypes.string.isRequired,
+  teaser: PropTypes.string,
+  body: PropTypes.string.isRequired,
   primaryImage: PropTypes.shape({
     src: PropTypes.string,
     focalPoint: PropTypes.shape({
